aider: Add assertion to check `battle_master_champion_id` equals 6 after deployment.
# Aider chat conversation:

USER: complete the todo

ASSISTANT: To complete the TODO in the test file, we need to add an assertion to check that `battle_master_champion_id` equals 6 after deployment.

Here are the *SEARCH/REPLACE* blocks:

test/BattleContract.js
```javascript
<<<<<<< SEARCH
    // TODO: test that battle_master_champion_id equals 6
=======
    expect(await battleContract.battle_master_champion_id()).to.equal(6);
>>>>>>> REPLACE
```

This will ensure that the test checks if the `battle_master_champion_id` is correctly set to 6 upon deployment.