aider: Ensure the battle master's image is twice as large as the challenger's image.
# Aider chat conversation:

USER: Okay, but make the size of the image only twice as large as the challengers image

ASSISTANT: To make the battle master's image twice as large as the challenger's image, we need to add some inline styles to the `img` tags in the `FightComponent`.

Here are the changes:

frontend/src/components/FightComponent.tsx
```typescript
<<<<<<< SEARCH
          <img src={challengerImage} alt="Your champion" />
=======
          <img src={challengerImage} alt="Your champion" style={{ width: '100px', height: '100px' }} />
>>>>>>> REPLACE
```

frontend/src/components/FightComponent.tsx
```typescript
<<<<<<< SEARCH
          <img src={battleMasterImage} alt="The Battlemaster" />
=======
          <img src={battleMasterImage} alt="The Battlemaster" style={{ width: '200px', height: '200px' }} />
>>>>>>> REPLACE
```

This will ensure that the battle master's image is twice as large as the challenger's image.